+++ src/main/java/controller/GameController.java
@@ -1,7 +1,7 @@
 package controller;
 
-import domain.GameManager;
-import domain.Player;
+import domain.game.GameManager;
+import domain.participant.Player;
 import domain.dto.GameInitialInfoDto;
 import view.InputView;
 import view.OutputView;
