+++ src/main/java/domain/DtoFactory.java
@@ -2,6 +2,9 @@ package domain;
 
 import domain.dto.GameInitialInfoDto;
 import domain.dto.GameScoreResultDto;
+import domain.participant.Dealer;
+import domain.participant.Player;
+import domain.participant.Players;
 import java.util.ArrayList;
 import java.util.List;
