+++ src/main/java/domain/Hand.java
@@ -1,6 +1,7 @@
 package domain;
 
-import domain.constant.Rank;
+import domain.card.Card;
+import domain.card.Rank;
 import java.util.ArrayList;
 import java.util.List;
