+++ src/test/java/domain/GameManagerTest.java
@@ -1,10 +1,11 @@
 package domain;
 
-import domain.constant.Rank;
-import domain.constant.Suit;
+import domain.card.Deck;
 import domain.dto.GameFinalResultDto;
 import domain.dto.GameInitialInfoDto;
 import domain.dto.GameScoreResultDto;
+import domain.game.GameManager;
+import domain.participant.Player;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
