+++ src/test/java/domain/GameResultJudgeTest.java
@@ -1,9 +1,14 @@
 package domain;
 
-import domain.constant.Rank;
+import domain.card.Card;
+import domain.card.Rank;
 import domain.constant.Result;
-import domain.constant.Suit;
+import domain.card.Suit;
 import domain.dto.GameFinalResultDto;
+import domain.game.GameResultJudge;
+import domain.participant.Dealer;
+import domain.participant.Player;
+import domain.participant.Players;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
